+++ README.md
@@ -4,9 +4,7 @@ A complete chat app that transcribes audio in real-time, streams back a response
 
 This repo is meant to serve as a starting point for your own language model-based apps, as well as a playground for experimentation. Contributions are welcome and encouraged!
 
-![quillman](https://user-images.githubusercontent.com/5786378/233804923-c13627de-97db-4050-a36b-62d955db9c19.gif)
-
-The language model used is [Zephyr](https://arxiv.org/abs/2310.16944). [OpenAI Whisper](https://github.com/openai/whisper) is used for transcription, and [Metavoice Tortoise TTS](https://github.com/metavoicexyz/tortoise-tts) is used for text-to-speech. The entire app, including the frontend, is made to be deployed serverlessly on [Modal](http://modal.com/).
+OpenAI [Whisper V3](https://huggingface.co/openai/whisper-large-v3) is used to produce a transcript, which is then passed into the [Zepyhr](https://arxiv.org/abs/2310.16944) language model to generate a response, which is then synthesized by Coqui's [XTTS](https://github.com/coqui-ai/TTS) text-to-speech model. All together, this produces a voice-to-voice chat experience.
 
 You can find the demo live [here](https://modal-labs--quillman-web.modal.run/).
 
@@ -16,9 +14,9 @@ You can find the demo live [here](https://modal-labs--quillman-web.modal.run/).
 
 1. React frontend ([`src/frontend/`](./src/frontend/))
 2. FastAPI server ([`src/app.py`](./src/app.py))
-3. Whisper transcription module ([`src/transcriber.py`](./src/transcriber.py))
-4. Tortoise text-to-speech module ([`src/tts.py`](./src/tts.py))
-5. Zephyr language model module ([`src/llm_zephyr.py`](./src/llm_zephyr.py))
+3. Whisper transcription module ([`src/whisper.py`](./src/whisper.py))
+4. XTTS text-to-speech module ([`src/xtts.py`](./src/xtts.py))
+5. Zephyr language model module ([`src/zephyr.py`](./src/zephyr.py))
 
 Read the accompanying [docs](https://modal.com/docs/examples/llm-voice-chat) for a detailed look at each of these components.
 
@@ -30,7 +28,21 @@ Read the accompanying [docs](https://modal.com/docs/examples/llm-voice-chat) for
 - A [Modal](http://modal.com/) account
 - A Modal token set up in your environment (`modal token new`)
 
-### Develop on Modal
+### Developing the inference modules
+
+Whisper, XTTS, and Zephyr each have a [local_entrypoint()](https://modal.com/docs/reference/modal.App#local_entrypoint) method that is invoked when you run that file directly. 
+This is useful for testing each module standalone, without needing to run the whole app.
+
+For example, to test the Whisper transcription module, run:
+```shell
+modal run -q src.whisper
+```
+
+### Developing the http server and frontend
+
+The http server at `src/app.py` is a [FastAPI](https://fastapi.tiangolo.com/) app that chains the inference modules into a single pipeline.
+
+It also serves the frontend as static files.
 
 To [serve](https://modal.com/docs/guide/webhooks#developing-with-modal-serve) the app on Modal, run this command from the root directory of this repo:
 
@@ -38,9 +50,9 @@ To [serve](https://modal.com/docs/guide/webhooks#developing-with-modal-serve) th
 modal serve src.app
 ```
 
-In the terminal output, you'll find a URL that you can visit to use your app. While the `modal serve` process is running, changes to any of the project files will be automatically applied. `Ctrl+C` will stop the app.
+In the terminal output, you'll find a URL that you can visit to use your app. While the `modal serve` process is running, changes to any of the project files will be automatically applied. `Ctrl+C` will stop the app. Note that for frontend changes, the browser cache will need to be cleared.
 
-### Deploy to Modal
+### Deploying to Modal
 
 Once you're happy with your changes, [deploy](https://modal.com/docs/guide/managing-deployments#creating-deployments) your app:
