--- src/common_proto.py
@@ -1,3 +1,0 @@
-from modal import App
-
-app = App(name="quillman-proto")
