--- src/frontend/spinners.css
@@ -1,194 +1,0 @@
-.talking {
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-  width: 13px;
-  height: 13px;
-}
-
-@keyframes bounce {
-  10% {
-    transform: scaleY(0.3);
-  }
-
-  30% {
-    transform: scaleY(1);
-  }
-
-  60% {
-    transform: scaleY(0.5);
-  }
-
-  80% {
-    transform: scaleY(0.75);
-  }
-
-  100% {
-    transform: scaleY(0.6);
-  }
-}
-
-.talking > span {
-  width: 3px;
-  height: 100%;
-  border-radius: 3px;
-  transform-origin: bottom;
-  animation: bounce 2.2s ease infinite alternate;
-  content: "";
-}
-
-.talking > span:nth-of-type(2) {
-  animation-delay: -2.2s;
-  /* Start at the end of animation */
-}
-
-.talking > span:nth-of-type(3) {
-  animation-delay: -3.7s;
-  /* Start mid-way of return of animation */
-}
-
-.silent {
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-  width: 13px;
-  height: 13px;
-}
-
-.silent > span {
-  width: 3px;
-  height: 100%;
-  border-radius: 3px;
-  transform-origin: bottom;
-  transform: scaleY(0.2);
-  content: "";
-}
-
-.lds-spinner {
-  color: official;
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-}
-
-.lds-spinner div {
-  /* transform-origin: 50% 50%; */
-  transform-origin: 40px 40px;
-  animation: lds-spinner 1.2s linear infinite;
-}
-
-.lds-spinner div:after {
-  content: " ";
-  display: block;
-  position: absolute;
-  top: 3px;
-  left: 37px;
-  width: 6px;
-  height: 18px;
-  border-radius: 20%;
-  /* background: #fff; */
-}
-
-.lds-spinner div:nth-child(1) {
-  transform: rotate(0deg);
-  animation-delay: -1.1s;
-}
-
-.lds-spinner div:nth-child(2) {
-  transform: rotate(30deg);
-  animation-delay: -1s;
-}
-
-.lds-spinner div:nth-child(3) {
-  transform: rotate(60deg);
-  animation-delay: -0.9s;
-}
-
-.lds-spinner div:nth-child(4) {
-  transform: rotate(90deg);
-  animation-delay: -0.8s;
-}
-
-.lds-spinner div:nth-child(5) {
-  transform: rotate(120deg);
-  animation-delay: -0.7s;
-}
-
-.lds-spinner div:nth-child(6) {
-  transform: rotate(150deg);
-  animation-delay: -0.6s;
-}
-
-.lds-spinner div:nth-child(7) {
-  transform: rotate(180deg);
-  animation-delay: -0.5s;
-}
-
-.lds-spinner div:nth-child(8) {
-  transform: rotate(210deg);
-  animation-delay: -0.4s;
-}
-
-.lds-spinner div:nth-child(9) {
-  transform: rotate(240deg);
-  animation-delay: -0.3s;
-}
-
-.lds-spinner div:nth-child(10) {
-  transform: rotate(270deg);
-  animation-delay: -0.2s;
-}
-
-.lds-spinner div:nth-child(11) {
-  transform: rotate(300deg);
-  animation-delay: -0.1s;
-}
-
-.lds-spinner div:nth-child(12) {
-  transform: rotate(330deg);
-  animation-delay: 0s;
-}
-
-@keyframes lds-spinner {
-  0% {
-    opacity: 1;
-  }
-
-  100% {
-    opacity: 0;
-  }
-}
-
-.gradient {
-  position: absolute;
-  width: 5000px;
-  height: 5000px;
-  left: -1400px;
-  top: -3200px;
-
-  background: radial-gradient(
-    rgba(51, 147, 112, 0.1) 10%,
-    rgba(232, 252, 14, 0.1) 25%,
-    rgba(20, 26, 18, 0.1) 45%,
-    rgba(20, 26, 18, 0) 50%
-  );
-}
-
-@keyframes pulse {
-  0% {
-    opacity: 1;
-  }
-
-  50% {
-    opacity: 0.7;
-  }
-
-  100% {
-    opacity: 1;
-  }
-}
-
-.pulse {
-  animation: pulse 2s ease-in-out infinite;
-}
